Generalize the Base Source for site changes. - Generalize Base Source to match layouts of most sources. - Update sub-sources to remove updated generalized functions.
