+++ src/Mangakakalot/Mangakakalot.ts
@@ -15,7 +15,8 @@ import { decodeHTML } from 'entities'
 
 import {
     MangaBox,
-    getExportVersion
+    getExportVersion,
+    HomeSectionsParams
 } from '../MangaBox'
 
 import { URLBuilder } from '../MangaBoxHelpers'
@@ -23,7 +24,7 @@ import { URLBuilder } from '../MangaBoxHelpers'
 const SITE_DOMAIN = 'https://mangakakalot.com'
 
 export const MangakakalotInfo: SourceInfo = {
-    version: getExportVersion('4.0.2'),
+    version: getExportVersion('3.1.2'),
     name: 'Mangakakalot',
     icon: 'icon.png',
     author: 'Batmeow',
@@ -32,7 +33,7 @@ export const MangakakalotInfo: SourceInfo = {
     contentRating: ContentRating.MATURE,
     websiteBaseURL: SITE_DOMAIN,
     sourceTags: [],
-    intents: SourceIntents.SETTINGS_UI | SourceIntents.HOMEPAGE_SECTIONS | SourceIntents.MANGA_CHAPTERS | SourceIntents.CLOUDFLARE_BYPASS_REQUIRED
+    intents: SourceIntents.SETTINGS_UI | SourceIntents.HOMEPAGE_SECTIONS | SourceIntents.MANGA_CHAPTERS
 }
 
 export class Mangakakalot extends MangaBox {
@@ -48,6 +49,12 @@ export class Mangakakalot extends MangaBox {
     // Appended path for manga list home sections.
     mangaListHomeSectionsPath = ''
 
+    // Homepage sections key value mappings.
+    override mangaListHomeSectionsParams: HomeSectionsParams = {
+        key: 'type',
+        values: ['latest', 'newest', 'topview']
+    }
+
     // Selector for manga in manga list.
     mangaListSelector = 'div.truyen-list div.list-truyen-item-wrap'
 
@@ -55,11 +62,7 @@ export class Mangakakalot extends MangaBox {
     mangaSubtitleSelector = 'a.list-story-item-wrap-chapter'
 
     // Page that requires captcha to access.
-    bypassPage = 'https://chapmanganato.to/'
-
-    override async supportsTagExclusion(): Promise<boolean> {
-        return false
-    }
+    bypassPage = ''
 
     override async getViewMoreItems(homePageSectionId: string, metadata: any): Promise<PagedResults> {
         const page: number = metadata?.page ?? 1
@@ -144,7 +147,7 @@ export class Mangakakalot extends MangaBox {
     }
     /* eslint-enable @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any */
 
-    parseTagId(url: string): string | undefined {
+    override parseTagId(url: string): string | undefined {
         return url.split('category=').pop()?.split('&')[0]?.replace(/[^0-9]/g, '')
     }
